pyproject.toml
[project]
name = 'catalogue'
version = '0.0.1'
dependencies = ["psycopg2-binary"]
